chat/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-jose[cryptography]==3.3.0
passlib[bcrypt]
jinja2
python-multipart
requests  # Добавьте эту строку
jwt
uvicorn[standard]
